requirements.txt
google-generativeai==0.4.0
colorama==0.4.6
Flask==3.0.0 

print("Available models for your API key:")
for m in genai.list_models():
    print(m.name) 